requirements.txt
streamlit
crewai
python-dotenv
PyPDF2
python-docx